pyproject.toml
[project]
name = "gc2-connect"
version = "0.1.0"
description = "Read Foresight GC2 launch monitor data and send to GSPro"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "nicegui>=1.4.0",
    "pyusb>=1.2.1",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
]

[project.scripts]
gc2-connect = "gc2_connect.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/gc2_connect"]

[tool.ruff]
line-length = 100
target-version = "py310"
